Add logical blocks and zookeeper stats monitor
- `DFS::get_metadata_optimized` function returns optimized logical
blocks.
- Added ls -o to display logical blocks in client
- Added runtime logical block generation
- IO Generator side works well and report to zk.
- Added several configuration variables in the json file.
- ZK_LISTENER is implemented, needs debugging.
- Fixed IO measure error
- Logical block scheduler Algorithm implemented
- Adding zookeeper support to travis CI

It is a squash merge with DFS master branch.

SQUASHED PARENT:
commit 28f5e3e20e2ca8694266ddf27539dbc49c8bb990
Merge: 6634654 b2ededc
Author: Vicente Adolfo Bolea Sanchez <vicente.bolea@gmail.com>
Date:   Mon Oct 23 18:00:48 2017 +0900

    Merge pull request #137 from DICL/cleaning-logical-blocks-feature

    Completes and isolates logical block feature

Conflicts:
	.github/PULL_REQUEST_TEMPLATE.md
	.travis.yml
	Makefile.am
	README.md
	configure.ac
	java_compile.sh
	set-env.sh
	src/client/cli_driver.cc
	src/client/cli_driver.hh
	src/client/dfs.cc
	src/client/dfs.hh
	src/client/model/metadata.hh
	src/client/vdfs.cc
	src/common/context.cc
	src/common/context_singleton.hh
	src/common/dl_loader.cc
	src/common/dl_loader.hh
	src/common/logger.cc
	src/common/settings.cc
	src/fileleader/directory.cc
	src/fileleader/file_leader.cc
	src/fileleader/file_leader.hh
	src/fileleader/file_leader_router.cc
	src/java/jni/velox_VeloxDFS.h
	src/java/velox/VeloxDFS.java
	src/java/velox/model/Metadata.java
	src/messages/blockupdate.hh
	src/messages/boost_impl.cc
	src/messages/boost_impl.hh
	src/messages/filedescription.hh
	src/messages/fileinfo.hh
	src/network/asyncchannel.cc
	src/network/client_handler.cc
	src/network/server_handler.cc
	src/targets/node_main.cc
	tests/Makefile.am
	tests/integration_test.sh

+++ src/mapreduce/task_executor.cc
@@ -163,7 +163,7 @@ void TaskExecutor::task_accept_status(TaskStatus* m) {
 void TaskExecutor::key_value_store(KeyValueShuffle *kv) {
   INFO("KVshuffle KV_ID=%lu, ID=%i, DST=%i", kv->kv_id, id, kv->node_id);
 
-  if (kv->node_id == id) {
+  if (kv->node_id == (uint32_t)id) {
 
     std::thread([&, this] (KeyValueShuffle kv) {
         try {
